requirements.txt
allure-pytest==2.9.44
allure-python-commons==2.9.44
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.5.30
charset-normalizer==2.0.6
chromedriver-autoinstaller==0.2.2
Faker==8.14.0
idna==3.2
iniconfig==1.1.1
mysql-connector-python==8.0.26
packaging==21.0
pluggy==1.0.0
protobuf==3.18.0
psycopg==3.0b1
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
requests==2.26.0
selenium==3.141.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
